start script fixes

+++ bin/tmcc_command_hub.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 source ~/venv/bin/activate
+export PYTHONPATH=~/BTTM/src/tmcc_command_hub
 
-python3 ~/BTTM/src/tmcc_command_hub/tmcc/dispatchers/serial_dispatcher.py &
-DISPATCHER_PID=$!
-
-python3 ~/BTTM/src/tmcc_command_hub/tmcc/web/app.py
-
-kill $DISPATCHER_PID
+if [ "$1" == "--app" ]; then
+    python3 ~/BTTM/src/tmcc_command_hub/tmcc/web/app.py
+else
+    python3 ~/BTTM/src/tmcc_command_hub/tmcc/dispatchers/serial_dispatcher.py &
+    DISPATCHER_PID=$!
+    python3 ~/BTTM/src/tmcc_command_hub/tmcc/web/app.py
+    kill $DISPATCHER_PID
+fi
